refactor: move request.js to a single package
