+++ packages/helper/src/index.js
@@ -1,4 +1,3 @@
-export { default as request } from './plugins/request';
 export { default as Render } from './Render/Render.vue';
 export { default as JSEditor } from './editor/JavaScriptEditor.vue';
 export { default as AtomicManager } from './manager/AtomicManager';
