+++ packages/request/src/index.js
@@ -1,0 +1,1 @@
+export { default as request } from './request';
